requirements_demo.txt
streamlit
grad-cam
opencv-python-headless
torch
torchvision
numpy
pillow
